in process

+++ src/pages/login/index.jsx
@@ -4,11 +4,10 @@ import s from './login.module.scss'
 import { useGetTokenMutation, useLoginMutation, useTokenRefreshMutation } from '../../services/user'
 import { useForm } from 'react-hook-form'
 import { useDispatch } from 'react-redux'
-import { setLogin } from '../../store/slices/user'
+import { setLogin, setLogout } from '../../store/slices/user'
 import { useEffect } from 'react'
 
 function LoginPage() {
-  // function LoginPage({ user, setUser }) {
   const navigate = useNavigate()
   const { register, handleSubmit } = useForm()
   const [login] = useLoginMutation()
@@ -16,46 +15,35 @@ function LoginPage() {
   const [tokenRefresh] = useTokenRefreshMutation()
   const dispatch = useDispatch()
 
-  // const [login, setLogin] = useState('')
-  // const [password, setPassword] = useState('')
+  const changeToken = async (token) => {
+    const responseRefresh = await tokenRefresh(token)
+    dispatch(changeToken({ ...responseRefresh.data }))
+  }
 
   useEffect(() => {
     const refresh = localStorage.getItem('refresh')
     if (!refresh) return
-
-    const changeToken = async () => {
-      try {
-        const responseRefresh = await tokenRefresh(refresh)
-        console.log(responseRefresh)
-        dispatch(changeToken({ ...responseRefresh.data }))
-        navigate('/')
-      } catch (error) {
-        // console.log(error?.detail)
-        localStorage.clear()
-      }
-    }
-
-    changeToken()
-    // navigate('/')
+    changeToken({ token: refresh })
   }, [])
 
-  // event.preventDefault()
-  // setUser('login', login, {
-  //   path: '/',
-  // })
-  // navigate('/', { replace: true })
   const onFormSubmit = async (fields) => {
-    // console.log(data)
-    try {
-      const responseLogin = await login({ ...fields })
-      const loginData = responseLogin.data
-      const responseToken = await getToken({ ...fields })
-      const tokenData = responseToken.data
-      dispatch(setLogin({ ...loginData, token: tokenData }))
-      localStorage.setItem('refresh', tokenData.refresh)
-    } catch (error) {
-      alert(error)
-    }
+    const responseLogin = await login({ ...fields })
+    const loginData = responseLogin.data
+
+    const responseToken = await getToken({ ...fields })
+    const tokenData = responseToken.data
+
+    dispatch(setLogin({ ...loginData, token: tokenData }))
+
+    localStorage.setItem('userID', loginData.id)
+    localStorage.setItem('refresh', tokenData.refresh)
+  }
+
+  if (useTokenRefreshMutation().isSuccess) navigate('/')
+
+  if (useTokenRefreshMutation().isError) {
+    setLogout()
+    localStorage.clear()
   }
 
   const handleRegistrationButtonClick = () => {
@@ -70,6 +58,10 @@ function LoginPage() {
           <div className={s.popup__fields}>
             <input placeholder="Логин" type="text" required className={s.popup__input} {...register('email')} />
             <input placeholder="Пароль" type="password" required className={s.popup__input} {...register('password')} />
+          </div>
+          <div className={s.popup__info}>
+            {useLoginMutation().isError && <div>Ошибка данных формы.</div>}
+            {useGetTokenMutation().isError && <div>Ошибка получения токена.</div>}
           </div>
           <div className={s.popup__buttons}>
             <button type="submit" className={s['button-primary']}>
